+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - `coinsort` now supports wildecards (\*) in recfolders, the parameter must be protected
 by single quote to avoid bash expantion 
 - `SubjectEndEP` and `SessionEndEP` plugins entry points for finalisation of session and subjects
+- Incorporated configuration file
+- logging options `level`, `quiet`, and `formatting` for log output control
+- save option for saving/updating locally configuration file
+- Search for map and config file in standard locations
 
 ### Fixed
 - EEG naming schema discrepency with BIDS standard
@@ -15,6 +19,8 @@ by single quote to avoid bash expantion
 - subject and session info are managed via BidsSession class
 - `participants.tsv` file now created at prepearing stage, with warning
 if field values are conflicting
+- Main code is `bidscoin.py` with cli command `prepare`, `bidsify` and `map`
+- Plugins exit code `< 0` do not throw exception  
 
 ## [dev2.1.0] - 2020-02-19